backend/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
channels==3.0.4
charset-normalizer==2.0.10
Django==4.0
django-ckeditor==6.2.0
django-js-asset==1.2.2
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
gunicorn==20.1.0
idna==3.3
Pillow==9.0.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
